Moving Engine methods into adapter.
