+++ .ci/unit-tests/SAP_Toolkit_Tests/SAP_Toolkit_Tests.csproj
@@ -16,4 +16,8 @@
     <PackageReference Include="coverlet.collector" Version="3.1.2" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="HelperMethods\" />
+  </ItemGroup>
+
 </Project>
