+++ SAP_Adapter/Convert/Argyle/OpeningType.cs
