+++ SAP_Adapter/Convert/Argyle/PropertyDetails.cs
