+++ SAP_Adapter/Convert/Argyle/SpecialFeatures.cs
