+++ SAP_Adapter/Convert/Argyle/ThermalBridge.cs
